requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
aiohttp==3.10.5
aiogram==3.13.1
aiohttp==3.10.5
python-dotenv==1.0.1
